Rename NoiseCapture package using noise-planet.org domain name
