+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "23.0.3"
 
     defaultConfig {
-        applicationId "org.orbisgis.protonomap"
+        applicationId "org.noise_planet.noisecapture"
         minSdkVersion 15
         targetSdkVersion 23
         versionCode 1
