+++ app/src/main/java/org/noise_planet/noisecapture/Map.java
@@ -1,9 +1,7 @@
-package org.orbisgis.protonomap;
+package org.noise_planet.noisecapture;
 
-import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
-import android.view.MenuItem;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.widget.ArrayAdapter;
