+++ app/src/main/java/org/noise_planet/noisecapture/MeasurementManager.java
@@ -1,4 +1,4 @@
-package org.orbisgis.protonomap;
+package org.noise_planet.noisecapture;
 
 import android.content.Context;
