+++ app/src/main/java/org/noise_planet/noisecapture/SPLValueFormatter.java
@@ -1,4 +1,4 @@
-package org.orbisgis.protonomap;
+package org.noise_planet.noisecapture;
 
 import com.github.mikephil.charting.utils.ValueFormatter;
