+++ app/src/main/java/org/noise_planet/noisecapture/Settings.java
@@ -1,12 +1,9 @@
-package org.orbisgis.protonomap;
+package org.noise_planet.noisecapture;
 
-import android.preference.ListPreference;
-import android.preference.Preference;
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.TextView;
 
 
 //
