+++ app/src/main/java/org/noise_planet/noisecapture/Start.java
@@ -1,8 +1,7 @@
-package org.orbisgis.protonomap;
+package org.noise_planet.noisecapture;
 
 import android.app.Activity;
 import android.content.Intent;
-import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
