+++ app/src/main/java/org/noise_planet/noisecapture/util/SystemUiHiderBase.java
@@ -1,4 +1,4 @@
-package org.orbisgis.protonomap.util;
+package org.noise_planet.noisecapture.util;
 
 import android.app.Activity;
 import android.view.View;
