+++ app/src/main/res/layout/activity_activity_calibration_start.xml
@@ -11,7 +11,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="org.orbisgis.protonomap.activity_calibration_start">
+    tools:context="org.noise_planet.noisecapture.activity_calibration_start">
 
         <TextView
             android:layout_width="wrap_content"
