+++ app/src/main/res/layout/activity_map.xml
@@ -11,7 +11,7 @@
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         android:paddingBottom="@dimen/activity_vertical_margin"
-        tools:context="org.orbisgis.protonomap.Map">
+        tools:context="org.noise_planet.noisecapture.Map">
 
         <LinearLayout
             android:orientation="vertical"
