+++ app/src/main/res/layout/activity_settings.xml
@@ -14,7 +14,7 @@
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         android:paddingBottom="@dimen/activity_vertical_margin"
-        tools:context="org.orbisgis.protonomap.Settings">
+        tools:context="org.noise_planet.noisecapture.Settings">
 
     </FrameLayout>
