+++ app/src/main/res/layout/activity_signal_processing.xml
@@ -1,4 +1,4 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:id="@+id/container"
     android:layout_width="match_parent" android:layout_height="match_parent"
-    tools:context="org.orbisgis.protonomap.SignalProcessing" tools:ignore="MergeRootFrame" />
+    tools:context="org.noise_planet.noisecapture.SignalProcessing" tools:ignore="MergeRootFrame" />
