+++ app/src/main/res/layout/activity_start.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="org.orbisgis.protonomap.Start"
+    tools:context="org.noise_planet.noisecapture.Start"
     android:background="#ffffffff"
     android:gravity="center_vertical|clip_vertical">
