+++ app/src/main/res/layout/activity_view_html_page.xml
@@ -11,7 +11,7 @@
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         android:paddingBottom="@dimen/activity_vertical_margin"
-        tools:context="org.orbisgis.protonomap.View_html_page">
+        tools:context="org.noise_planet.noisecapture.View_html_page">
 
         <TextView android:text="@string/hello_world" android:layout_width="wrap_content"
             android:layout_height="wrap_content" />
