+++ app/src/main/res/layout/fragment_signal_processing.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="org.orbisgis.protonomap.SignalProcessing$PlaceholderFragment">
+    tools:context="org.noise_planet.noisecapture.SignalProcessing$PlaceholderFragment">
 
     <TextView android:text="@string/hello_world" android:layout_width="wrap_content"
         android:layout_height="wrap_content"
