+++ app/src/main/res/menu/menu_about.xml
@@ -1,5 +1,5 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="org.orbisgis.protonomap.MainActivity">
+    tools:context="org.noise_planet.noisecapture.MainActivity">
 </menu>
