+++ app/src/main/res/menu/menu_start.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools" tools:context="org.orbisgis.protonomap.Start">
+    xmlns:tools="http://schemas.android.com/tools" tools:context="org.noise_planet.noisecapture.Start">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" app:showAsAction="never" />
 </menu>
